Why different dimensions in different sections??
On line#45, collection `test01` is set to 256 dimension.
On line#142, vectors of 256 dimensions are inserted into `test01`
However, on line#227, docs are saying to create query vector of "32-dimension". This will surely run into error: (Status(code=7, message='The vector dimension must be equal to the collection dimension.').

Nearest neighbor search algorithms assume FIXED dimension of dataset, inserted feature vectors and query vectors.
So, as in my previous pull request, I have proposed to set 'dim' parameter in the top of docs, when collection was created. 
OR instead of using `dim`, writed down direct values of the dimensions which will match your coding style used in this page.